Expand retry feature, refactor and improvements
