+++ src/autostake/Health.mjs
@@ -1,51 +1,51 @@
 import _ from 'lodash'
 import axios from 'axios'
-import { timeStamp } from './Helpers.mjs'
+import { timeStamp } from '../utils/Helpers.mjs'
 
-class AutostakeHealth {
+class Health {
   constructor(config, opts) {
     const { address, uuid } = config || {}
     const { dryRun } = opts || {}
     this.address = address || 'https://hc-ping.com'
     this.uuid = uuid
     this.dryRun = dryRun
+    this.logs = []
   }
 
   started(...args){
     timeStamp(...args)
     if(this.uuid) timeStamp('Starting health', [this.address, this.uuid].join('/'))
-    this.ping('start', ...args)
-  }
-
-  error(...args){
-    timeStamp(...args)
-    this.errors = [...(this.errors || []), "\n", ...args]
-  }
-
-  complete(...args){
-    timeStamp(...args)
-    if(this.errors){
-      this.ping('fail', ...[...this.errors, "\n", ...args])
-    }else{
-      this.ping(undefined, ...args)
-    }
+    return this.ping('start', ...args)
   }
 
   success(...args){
     timeStamp(...args)
-    this.ping(undefined, ...args)
+    return this.ping(undefined, ...[...this.logs, "\n", ...args])
   }
 
   failed(...args){
     timeStamp(...args)
-    this.ping('fail', ...args)
+    return this.ping('fail', ...[...this.logs, "\n", ...args])
+  }
+
+  log(...args){
+    timeStamp(...args)
+    this.logs = [...this.logs, "\n", ...args]
+  }
+
+  addLogs(logs){
+    this.logs.concat(logs.map(el => [el, "\n"]).flat())
+  }
+
+  sendLog(){
+    return this.ping('log', ...this.logs)
   }
 
   ping(action, ...args){
     if(!this.uuid) return
     if(this.dryRun) return timeStamp('DRYRUN: Skipping health check ping')
 
-    axios.request({
+    return axios.request({
       method: 'POST',
       url: _.compact([this.address, this.uuid, action]).join('/'), 
       data: args.join(' ')
@@ -55,4 +55,4 @@ class AutostakeHealth {
   }
 }
 
-export default AutostakeHealth
+export default Health
